Implemented the number of extra images and orientation to display them for images template fields (previously used for the entry images only). Horizontally-displayed images are not fully implemented (displays off-screen if width exceeds visible window width).
Template deletion is implemented as red text on the bottom right of the template editor dialog. It clears the template and associated collections / data.

The entry images field can no longer be deleted from the second column of the template field editor.

Implemented filtered searching with template fields on the main gui.

Multiple images can be uploaded at once with any entry images or images field.

Searchbox is keyboard-navigable and responds to clicking and pressing enter. As a result, tabbing to it and using keyboard no longer throws an InvalidOperationException. RemoveDiacritics() is now a private method.